Commit April 24, 2020
The app takes pictures and stores and retrieves them from db. I added 3 fields family table in db: one for setting it to emergency contact, one for image, one for whether checkbox is checked for birthdays. The Emergency activity now loads emergency contacts, the all family members activity now includes an image for each family member or the placeholder thumbnail if none was taken. 
Current Problem --> the app crashes whenever I edit the image after saving the family member... emergency button is not displaying properly, App needs to call natively from itself(in best case scenario, i think, if it is going to include a 911 button... im not sure yet though... I want to have the first version in by the end of this week.